% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_aslist.R
\name{read_aslist}
\alias{read_aslist}
\title{Make a list file from large character file}
\usage{
read_aslist(id = id, all_text = all_text, con = con)
}
\arguments{
\item{id}{The id list made from sub_fasID function.}

\item{all_text}{Large character read in by readLines function from fasta file.}

\item{con}{A connection object or a character string, the connection must refer to the same fasta file as all_text.}
}
\value{
Chromosome fasta file in list format.
}
\description{
Make a list file from large character file
}
\examples{
data("all_text")
id <- sub_fasID(all_text = all_text)
fil <- tempfile(fileext = ".data")
write(all_text,file = fil)
con0 <- file(fil, "r")
tex <- read_aslist(id,all_text = all_text,con = con0)
}
