% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_list.R
\name{sort_list}
\alias{sort_list}
\title{Sort the chromosome list  according to the chromosome number}
\usage{
sort_list(id = id, tex = tex, chrsig = "single")
}
\arguments{
\item{id}{The identifier list of the fasta file made by sub_fasID.}

\item{tex}{A chromosome fasta file in list format made by read_aslist}

\item{chrsig}{The number of characters of the chromosome. Either "single"(means a single character following "chr" in the fasta identifier) or "double"(means two characters following "chr" in the fasta identifier). eg."chr1,chrX,chrY,chrM" is "single";"chr10,chr11" is "double".
If you want to obtain both "single" and "double" sorted list of chromosome, try "single" and "double" respectively.}
}
\value{
The sorted chromosome fasta file in list format
}
\description{
Sort the chromosome list  according to the chromosome number
}
\examples{
data("tex")
data("all_text")
all_text<- replace_all_text(input_type = "all_text",input = all_text)
id <- sub_fasID(all_text = all_text)
tex2<- sort_list(id=id,tex = tex,chrsig = "single")
tex3 <- sort_list(id=id,tex = tex,chrsig = "double")
}
