% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitChr.R
\name{splitChr}
\alias{splitChr}
\title{Split all chromosomes from the sorted chromosome list}
\usage{
splitChr(tex = tex2, chr = chr, sex = FALSE)
}
\arguments{
\item{tex}{The sorted chromosome list made by sort_list.}

\item{chr}{The chromosome sequence,if the chromosome list is "single", try seq(1,9) because you should start with 1; if the chromosome list is "double", try seq(10,22) because you should start with 10.}

\item{sex}{Whether to output the sex chromosomes like X chromosome and Y chromosome.}
}
\value{
Write the splitted chromosome fasta file to separated txt file according to the chromosome number.
}
\description{
Split all chromosomes from the sorted chromosome list
}
\examples{
data("tex2")
data("tex3")
splitChr(tex = tex2,chr=seq(1,9),sex = TRUE)
#chromosome X or Y is "single",so the tex should be a
#"single" chromosome list.
#In the below case, sex should be F.
splitChr(tex = tex3,chr=seq(10,22),sex = FALSE)
}
