% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex.R
\docType{data}
\name{tex}
\alias{tex}
\title{Fasta file of chromosome sequence produced from all_text}
\format{A large list containing 411 elements}
\usage{
tex
}
\description{
A dataset containing the hg19 chromosome sequence from the
hg19 blacklist.
}
\details{
Please first refer to the document of all_text dataset.
The all_text dataset is transformed into tex by using
read_aslist function.
}
\references{
\url{https://doi.org/10.1038/s41587-019-0206-z}
}
\keyword{datasets}
