% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex3.R
\docType{data}
\name{tex3}
\alias{tex3}
\title{The second part of fasta file of chromosome sequence produced
from tex}
\format{A large list containing 106 elements.}
\usage{
tex3
}
\description{
A dataset containing the hg19 chromosome sequence from the
hg19 blacklist, but only chr10 to chr22.
}
\details{
Please first refer to the document of all_text dataset
and tex dataset.
The tex dataset is transformed into tex3 by using sort_list
function:sort_list(id=id,tex = tex,chrsig = "double")
}
\references{
\url{https://doi.org/10.1038/s41587-019-0206-z}
}
\keyword{datasets}
