% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complicated_id.R
\docType{data}
\name{complicated_id}
\alias{complicated_id}
\title{Sampled Fasta file of chromosome sequence from hg19 blacklist}
\format{A character sequence with 20 elements}
\usage{
complicated_id
}
\description{
This dataset is sampled from The hg19 blacklist.
}
\details{
For splitting a chromosome fasta file, sometimes the
fasta identifier is too complicated to manipulate.This data
can be used to show how to simplify the fasta identifier.
}
\references{
\url{https://doi.org/10.1038/s41587-019-0206-z}
}
\keyword{datasets}
