% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_aslist.R
\name{read_aslist}
\alias{read_aslist}
\title{Make a list file from large character file}
\usage{
read_aslist(id = id, all_text = all_text)
}
\arguments{
\item{id}{The id list made from sub_fasID function.}

\item{all_text}{Large character read by readLines function from fasta file.}
}
\value{
A chromosome list file
}
\description{
Make a list file from large character file
}
\examples{
data("all_text")
id <- sub_fasID(all_text = all_text)
tex <- read_aslist(id,all_text = all_text)
}
