% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_all_text.R
\name{replace_all_text}
\alias{replace_all_text}
\title{Replace tedious chromosome identifier into simple format}
\usage{
replace_all_text(input_type = "all_text", input = input)
}
\arguments{
\item{input_type}{This can be either "all_text" or "list", The previous is a large character containing each line of the fasta file, the latter is a list in which each element contains a unit of fasta file.}

\item{input}{The large character or list containing ids that need to be simplified.}
}
\value{
The large character or list of Chromosome fasta file with simplified id
}
\description{
Make the chromosome id starting with ">" into simple format like:
">chr:1091194-1093520...",this is helpful for sorting the chromosome according to their number
}
\examples{
data("complicated_id")
all_text<- replace_all_text(input_type = "all_text",input = complicated_id)
}
