% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitChr.R
\name{splitChr}
\alias{splitChr}
\title{Split all chromosomes from the sorted chromosome list}
\usage{
splitChr(tex = tex, chr = chr, sex = FALSE)
}
\arguments{
\item{tex}{The sorted chromosome list made by sort_list.}

\item{chr}{The chromosome number sequence,if the chromosome list is "single"(means a single character following "chr" in the fasta identifier), try seq(1,9), be sure that starting with 1; if the chromosome list is "double"(means two characters following "chr" in the fasta identifier), try seq(10,22), be sure that starting with 10 but the ending can be changed.}

\item{sex}{Whether to output the sex chromosomes like X chromosome and Y chromosome.}
}
\value{
Write the splitted chromosome fasta file to separated txt files according to the chromosome number.
}
\description{
Split all chromosomes from the sorted chromosome list
}
\examples{
data("tex")
data("all_text")
all_text<- replace_all_text(input_type = "all_text",input = all_text)
id <- sub_fasID(all_text = all_text)
tex2<- sort_list(id=id,tex = tex,chrsig = "single")
tex3 <- sort_list(id=id,tex = tex,chrsig = "double")
splitChr(tex = tex2,chr=seq(1,9),sex = TRUE)
#chromosome X or Y is "single",so the tex should be a
#"single" chromosome list.
#In the below case, sex should be F.
splitChr(tex = tex3,chr=seq(10,22),sex = FALSE)
}
