% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_fasID.R
\name{sub_fasID}
\alias{sub_fasID}
\title{Subtract chromosome ids from fasta file}
\usage{
sub_fasID(all_text = all_text)
}
\arguments{
\item{all_text}{Large character read by readLines from chromosome fasta file.}
}
\value{
The id list of the fasta file(fasta identifier), refer to http://genetics.bwh.harvard.edu/pph/FASTA.html
}
\description{
Subtract chromosome ids from fasta file
}
\examples{
data("all_text")
all_text<- replace_all_text(input_type = "all_text",input = all_text)
id <- sub_fasID(all_text = all_text)
}
