% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tex}
\alias{tex}
\title{Fasta file of chromosome sequence produced from text dataset}
\format{A large list containing 62 elements}
\usage{
data(tex)
}
\description{
A downsampled dataset containing the hg19 chromosome sequence
from the hg19 blacklist. Please first refer to the document of
text dataset.
For splitting the text, please first transformed text
into a list in which each element contains a chromosome fasta
file. eg.The text dataset is transformed into tex.
}
\examples{
data(tex)
}
\references{
Satpathy A T, Granja J M, Yost K E, et al. (2019) Nature biotechnology 37,925–936.
(\href{https://www.ncbi.nlm.nih.gov/pubmed/31375813}{PubMed})
}
\keyword{datasets}
